fix: ウェザーリポート一覧取得用のDtoをapi仕様に修正
