+++ reimi/frontend/reimi_app/lib/data/dtos/weather_report_dto.freezed.dart
@@ -11,25 +11,19 @@ part of 'weather_report_dto.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
+WeatherReportDto _$WeatherReportDtoFromJson(Map<String, dynamic> json) {
+  return _WeatherReportSimpleDto.fromJson(json);
+}
 
 /// @nodoc
 mixin _$WeatherReportDto {
-  String get reportId;
+  String get id;
   String get userId;
-  String get userName;
-  String get mainPhotoUrl;
   String get comment;
-  WeatherType get weatherType;
-  FeelingType get feelingType;
-  ForecastType get forecastType;
   MediaType get mediaType;
   String get url;
+  @IsoDateTimeConverter()
   DateTime get createdAt;
-  List<String>? get reportComment;
-  int? get likeCount;
-  int? get commentCount;
-  double? get latitude;
-  double? get longitude;
 
   /// Create a copy of WeatherReportDto
   /// with the given fields replaced by the non-null parameter values.
@@ -47,61 +41,24 @@ mixin _$WeatherReportDto {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is WeatherReportDto &&
-            (identical(other.reportId, reportId) ||
-                other.reportId == reportId) &&
+            (identical(other.id, id) || other.id == id) &&
             (identical(other.userId, userId) || other.userId == userId) &&
-            (identical(other.userName, userName) ||
-                other.userName == userName) &&
-            (identical(other.mainPhotoUrl, mainPhotoUrl) ||
-                other.mainPhotoUrl == mainPhotoUrl) &&
             (identical(other.comment, comment) || other.comment == comment) &&
-            (identical(other.weatherType, weatherType) ||
-                other.weatherType == weatherType) &&
-            (identical(other.feelingType, feelingType) ||
-                other.feelingType == feelingType) &&
-            (identical(other.forecastType, forecastType) ||
-                other.forecastType == forecastType) &&
             (identical(other.mediaType, mediaType) ||
                 other.mediaType == mediaType) &&
             (identical(other.url, url) || other.url == url) &&
             (identical(other.createdAt, createdAt) ||
-                other.createdAt == createdAt) &&
-            const DeepCollectionEquality()
-                .equals(other.reportComment, reportComment) &&
-            (identical(other.likeCount, likeCount) ||
-                other.likeCount == likeCount) &&
-            (identical(other.commentCount, commentCount) ||
-                other.commentCount == commentCount) &&
-            (identical(other.latitude, latitude) ||
-                other.latitude == latitude) &&
-            (identical(other.longitude, longitude) ||
-                other.longitude == longitude));
+                other.createdAt == createdAt));
   }
 
   @JsonKey(includeFromJson: false, includeToJson: false)
   @override
-  int get hashCode => Object.hash(
-      runtimeType,
-      reportId,
-      userId,
-      userName,
-      mainPhotoUrl,
-      comment,
-      weatherType,
-      feelingType,
-      forecastType,
-      mediaType,
-      url,
-      createdAt,
-      const DeepCollectionEquality().hash(reportComment),
-      likeCount,
-      commentCount,
-      latitude,
-      longitude);
+  int get hashCode =>
+      Object.hash(runtimeType, id, userId, comment, mediaType, url, createdAt);
 
   @override
   String toString() {
-    return 'WeatherReportDto(reportId: $reportId, userId: $userId, userName: $userName, mainPhotoUrl: $mainPhotoUrl, comment: $comment, weatherType: $weatherType, feelingType: $feelingType, forecastType: $forecastType, mediaType: $mediaType, url: $url, createdAt: $createdAt, reportComment: $reportComment, likeCount: $likeCount, commentCount: $commentCount, latitude: $latitude, longitude: $longitude)';
+    return 'WeatherReportDto(id: $id, userId: $userId, comment: $comment, mediaType: $mediaType, url: $url, createdAt: $createdAt)';
   }
 }
 
@@ -112,22 +69,12 @@ abstract mixin class $WeatherReportDtoCopyWith<$Res> {
       _$WeatherReportDtoCopyWithImpl;
   @useResult
   $Res call(
-      {String reportId,
+      {String id,
       String userId,
-      String userName,
-      String mainPhotoUrl,
       String comment,
-      WeatherType weatherType,
-      FeelingType feelingType,
-      ForecastType forecastType,
       MediaType mediaType,
       String url,
-      DateTime createdAt,
-      List<String>? reportComment,
-      int? likeCount,
-      int? commentCount,
-      double? latitude,
-      double? longitude});
+      @IsoDateTimeConverter() DateTime createdAt});
 }
 
 /// @nodoc
@@ -143,56 +90,26 @@ class _$WeatherReportDtoCopyWithImpl<$Res>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
-    Object? reportId = null,
+    Object? id = null,
     Object? userId = null,
-    Object? userName = null,
-    Object? mainPhotoUrl = null,
     Object? comment = null,
-    Object? weatherType = null,
-    Object? feelingType = null,
-    Object? forecastType = null,
     Object? mediaType = null,
     Object? url = null,
     Object? createdAt = null,
-    Object? reportComment = freezed,
-    Object? likeCount = freezed,
-    Object? commentCount = freezed,
-    Object? latitude = freezed,
-    Object? longitude = freezed,
   }) {
     return _then(_self.copyWith(
-      reportId: null == reportId
-          ? _self.reportId
-          : reportId // ignore: cast_nullable_to_non_nullable
+      id: null == id
+          ? _self.id
+          : id // ignore: cast_nullable_to_non_nullable
               as String,
       userId: null == userId
           ? _self.userId
           : userId // ignore: cast_nullable_to_non_nullable
               as String,
-      userName: null == userName
-          ? _self.userName
-          : userName // ignore: cast_nullable_to_non_nullable
-              as String,
-      mainPhotoUrl: null == mainPhotoUrl
-          ? _self.mainPhotoUrl
-          : mainPhotoUrl // ignore: cast_nullable_to_non_nullable
-              as String,
       comment: null == comment
           ? _self.comment
           : comment // ignore: cast_nullable_to_non_nullable
               as String,
-      weatherType: null == weatherType
-          ? _self.weatherType
-          : weatherType // ignore: cast_nullable_to_non_nullable
-              as WeatherType,
-      feelingType: null == feelingType
-          ? _self.feelingType
-          : feelingType // ignore: cast_nullable_to_non_nullable
-              as FeelingType,
-      forecastType: null == forecastType
-          ? _self.forecastType
-          : forecastType // ignore: cast_nullable_to_non_nullable
-              as ForecastType,
       mediaType: null == mediaType
           ? _self.mediaType
           : mediaType // ignore: cast_nullable_to_non_nullable
@@ -205,26 +122,6 @@ class _$WeatherReportDtoCopyWithImpl<$Res>
           ? _self.createdAt
           : createdAt // ignore: cast_nullable_to_non_nullable
               as DateTime,
-      reportComment: freezed == reportComment
-          ? _self.reportComment
-          : reportComment // ignore: cast_nullable_to_non_nullable
-              as List<String>?,
-      likeCount: freezed == likeCount
-          ? _self.likeCount
-          : likeCount // ignore: cast_nullable_to_non_nullable
-              as int?,
-      commentCount: freezed == commentCount
-          ? _self.commentCount
-          : commentCount // ignore: cast_nullable_to_non_nullable
-              as int?,
-      latitude: freezed == latitude
-          ? _self.latitude
-          : latitude // ignore: cast_nullable_to_non_nullable
-              as double?,
-      longitude: freezed == longitude
-          ? _self.longitude
-          : longitude // ignore: cast_nullable_to_non_nullable
-              as double?,
     ));
   }
 }
@@ -245,12 +142,12 @@ extension WeatherReportDtoPatterns on WeatherReportDto {
 
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>(
-    TResult Function(_WeatherReportDto value)? $default, {
+    TResult Function(_WeatherReportSimpleDto value)? $default, {
     required TResult orElse(),
   }) {
     final _that = this;
     switch (_that) {
-      case _WeatherReportDto() when $default != null:
+      case _WeatherReportSimpleDto() when $default != null:
         return $default(_that);
       case _:
         return orElse();
@@ -272,11 +169,11 @@ extension WeatherReportDtoPatterns on WeatherReportDto {
 
   @optionalTypeArgs
   TResult map<TResult extends Object?>(
-    TResult Function(_WeatherReportDto value) $default,
+    TResult Function(_WeatherReportSimpleDto value) $default,
   ) {
     final _that = this;
     switch (_that) {
-      case _WeatherReportDto():
+      case _WeatherReportSimpleDto():
         return $default(_that);
       case _:
         throw StateError('Unexpected subclass');
@@ -297,11 +194,11 @@ extension WeatherReportDtoPatterns on WeatherReportDto {
 
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_WeatherReportDto value)? $default,
+    TResult? Function(_WeatherReportSimpleDto value)? $default,
   ) {
     final _that = this;
     switch (_that) {
-      case _WeatherReportDto() when $default != null:
+      case _WeatherReportSimpleDto() when $default != null:
         return $default(_that);
       case _:
         return null;
@@ -323,45 +220,20 @@ extension WeatherReportDtoPatterns on WeatherReportDto {
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>(
     TResult Function(
-            String reportId,
+            String id,
             String userId,
-            String userName,
-            String mainPhotoUrl,
             String comment,
-            WeatherType weatherType,
-            FeelingType feelingType,
-            ForecastType forecastType,
             MediaType mediaType,
             String url,
-            DateTime createdAt,
-            List<String>? reportComment,
-            int? likeCount,
-            int? commentCount,
-            double? latitude,
-            double? longitude)?
+            @IsoDateTimeConverter() DateTime createdAt)?
         $default, {
     required TResult orElse(),
   }) {
     final _that = this;
     switch (_that) {
-      case _WeatherReportDto() when $default != null:
-        return $default(
-            _that.reportId,
-            _that.userId,
-            _that.userName,
-            _that.mainPhotoUrl,
-            _that.comment,
-            _that.weatherType,
-            _that.feelingType,
-            _that.forecastType,
-            _that.mediaType,
-            _that.url,
-            _that.createdAt,
-            _that.reportComment,
-            _that.likeCount,
-            _that.commentCount,
-            _that.latitude,
-            _that.longitude);
+      case _WeatherReportSimpleDto() when $default != null:
+        return $default(_that.id, _that.userId, _that.comment, _that.mediaType,
+            _that.url, _that.createdAt);
       case _:
         return orElse();
     }
@@ -383,44 +255,19 @@ extension WeatherReportDtoPatterns on WeatherReportDto {
   @optionalTypeArgs
   TResult when<TResult extends Object?>(
     TResult Function(
-            String reportId,
+            String id,
             String userId,
-            String userName,
-            String mainPhotoUrl,
             String comment,
-            WeatherType weatherType,
-            FeelingType feelingType,
-            ForecastType forecastType,
             MediaType mediaType,
             String url,
-            DateTime createdAt,
-            List<String>? reportComment,
-            int? likeCount,
-            int? commentCount,
-            double? latitude,
-            double? longitude)
+            @IsoDateTimeConverter() DateTime createdAt)
         $default,
   ) {
     final _that = this;
     switch (_that) {
-      case _WeatherReportDto():
-        return $default(
-            _that.reportId,
-            _that.userId,
-            _that.userName,
-            _that.mainPhotoUrl,
-            _that.comment,
-            _that.weatherType,
-            _that.feelingType,
-            _that.forecastType,
-            _that.mediaType,
-            _that.url,
-            _that.createdAt,
-            _that.reportComment,
-            _that.likeCount,
-            _that.commentCount,
-            _that.latitude,
-            _that.longitude);
+      case _WeatherReportSimpleDto():
+        return $default(_that.id, _that.userId, _that.comment, _that.mediaType,
+            _that.url, _that.createdAt);
       case _:
         throw StateError('Unexpected subclass');
     }
@@ -441,44 +288,19 @@ extension WeatherReportDtoPatterns on WeatherReportDto {
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>(
     TResult? Function(
-            String reportId,
+            String id,
             String userId,
-            String userName,
-            String mainPhotoUrl,
             String comment,
-            WeatherType weatherType,
-            FeelingType feelingType,
-            ForecastType forecastType,
             MediaType mediaType,
             String url,
-            DateTime createdAt,
-            List<String>? reportComment,
-            int? likeCount,
-            int? commentCount,
-            double? latitude,
-            double? longitude)?
+            @IsoDateTimeConverter() DateTime createdAt)?
         $default,
   ) {
     final _that = this;
     switch (_that) {
-      case _WeatherReportDto() when $default != null:
-        return $default(
-            _that.reportId,
-            _that.userId,
-            _that.userName,
-            _that.mainPhotoUrl,
-            _that.comment,
-            _that.weatherType,
-            _that.feelingType,
-            _that.forecastType,
-            _that.mediaType,
-            _that.url,
-            _that.createdAt,
-            _that.reportComment,
-            _that.likeCount,
-            _that.commentCount,
-            _that.latitude,
-            _that.longitude);
+      case _WeatherReportSimpleDto() when $default != null:
+        return $default(_that.id, _that.userId, _that.comment, _that.mediaType,
+            _that.url, _that.createdAt);
       case _:
         return null;
     }
@@ -487,80 +309,43 @@ extension WeatherReportDtoPatterns on WeatherReportDto {
 
 /// @nodoc
 @JsonSerializable()
-class _WeatherReportDto implements WeatherReportDto {
-  const _WeatherReportDto(
-      {required this.reportId,
+class _WeatherReportSimpleDto implements WeatherReportDto {
+  const _WeatherReportSimpleDto(
+      {required this.id,
       required this.userId,
-      required this.userName,
-      required this.mainPhotoUrl,
       required this.comment,
-      required this.weatherType,
-      required this.feelingType,
-      required this.forecastType,
       required this.mediaType,
       required this.url,
-      required this.createdAt,
-      final List<String>? reportComment,
-      this.likeCount,
-      this.commentCount,
-      this.latitude,
-      this.longitude})
-      : _reportComment = reportComment;
-  factory _WeatherReportDto.fromJson(Map<String, dynamic> json) =>
-      _$WeatherReportDtoFromJson(json);
+      @IsoDateTimeConverter() required this.createdAt});
+  factory _WeatherReportSimpleDto.fromJson(Map<String, dynamic> json) =>
+      _$WeatherReportSimpleDtoFromJson(json);
 
   @override
-  final String reportId;
+  final String id;
   @override
   final String userId;
   @override
-  final String userName;
-  @override
-  final String mainPhotoUrl;
-  @override
   final String comment;
-  @override
-  final WeatherType weatherType;
-  @override
-  final FeelingType feelingType;
-  @override
-  final ForecastType forecastType;
   @override
   final MediaType mediaType;
   @override
   final String url;
   @override
+  @IsoDateTimeConverter()
   final DateTime createdAt;
-  final List<String>? _reportComment;
-  @override
-  List<String>? get reportComment {
-    final value = _reportComment;
-    if (value == null) return null;
-    if (_reportComment is EqualUnmodifiableListView) return _reportComment;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(value);
-  }
-
-  @override
-  final int? likeCount;
-  @override
-  final int? commentCount;
-  @override
-  final double? latitude;
-  @override
-  final double? longitude;
 
   /// Create a copy of WeatherReportDto
   /// with the given fields replaced by the non-null parameter values.
   @override
   @JsonKey(includeFromJson: false, includeToJson: false)
   @pragma('vm:prefer-inline')
-  _$WeatherReportDtoCopyWith<_WeatherReportDto> get copyWith =>
-      __$WeatherReportDtoCopyWithImpl<_WeatherReportDto>(this, _$identity);
+  _$WeatherReportSimpleDtoCopyWith<_WeatherReportSimpleDto> get copyWith =>
+      __$WeatherReportSimpleDtoCopyWithImpl<_WeatherReportSimpleDto>(
+          this, _$identity);
 
   @override
   Map<String, dynamic> toJson() {
-    return _$WeatherReportDtoToJson(
+    return _$WeatherReportSimpleDtoToJson(
       this,
     );
   }
@@ -569,155 +354,78 @@ class _WeatherReportDto implements WeatherReportDto {
   bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _WeatherReportDto &&
-            (identical(other.reportId, reportId) ||
-                other.reportId == reportId) &&
+            other is _WeatherReportSimpleDto &&
+            (identical(other.id, id) || other.id == id) &&
             (identical(other.userId, userId) || other.userId == userId) &&
-            (identical(other.userName, userName) ||
-                other.userName == userName) &&
-            (identical(other.mainPhotoUrl, mainPhotoUrl) ||
-                other.mainPhotoUrl == mainPhotoUrl) &&
             (identical(other.comment, comment) || other.comment == comment) &&
-            (identical(other.weatherType, weatherType) ||
-                other.weatherType == weatherType) &&
-            (identical(other.feelingType, feelingType) ||
-                other.feelingType == feelingType) &&
-            (identical(other.forecastType, forecastType) ||
-                other.forecastType == forecastType) &&
             (identical(other.mediaType, mediaType) ||
                 other.mediaType == mediaType) &&
             (identical(other.url, url) || other.url == url) &&
             (identical(other.createdAt, createdAt) ||
-                other.createdAt == createdAt) &&
-            const DeepCollectionEquality()
-                .equals(other._reportComment, _reportComment) &&
-            (identical(other.likeCount, likeCount) ||
-                other.likeCount == likeCount) &&
-            (identical(other.commentCount, commentCount) ||
-                other.commentCount == commentCount) &&
-            (identical(other.latitude, latitude) ||
-                other.latitude == latitude) &&
-            (identical(other.longitude, longitude) ||
-                other.longitude == longitude));
+                other.createdAt == createdAt));
   }
 
   @JsonKey(includeFromJson: false, includeToJson: false)
   @override
-  int get hashCode => Object.hash(
-      runtimeType,
-      reportId,
-      userId,
-      userName,
-      mainPhotoUrl,
-      comment,
-      weatherType,
-      feelingType,
-      forecastType,
-      mediaType,
-      url,
-      createdAt,
-      const DeepCollectionEquality().hash(_reportComment),
-      likeCount,
-      commentCount,
-      latitude,
-      longitude);
+  int get hashCode =>
+      Object.hash(runtimeType, id, userId, comment, mediaType, url, createdAt);
 
   @override
   String toString() {
-    return 'WeatherReportDto(reportId: $reportId, userId: $userId, userName: $userName, mainPhotoUrl: $mainPhotoUrl, comment: $comment, weatherType: $weatherType, feelingType: $feelingType, forecastType: $forecastType, mediaType: $mediaType, url: $url, createdAt: $createdAt, reportComment: $reportComment, likeCount: $likeCount, commentCount: $commentCount, latitude: $latitude, longitude: $longitude)';
+    return 'WeatherReportDto(id: $id, userId: $userId, comment: $comment, mediaType: $mediaType, url: $url, createdAt: $createdAt)';
   }
 }
 
 /// @nodoc
-abstract mixin class _$WeatherReportDtoCopyWith<$Res>
+abstract mixin class _$WeatherReportSimpleDtoCopyWith<$Res>
     implements $WeatherReportDtoCopyWith<$Res> {
-  factory _$WeatherReportDtoCopyWith(
-          _WeatherReportDto value, $Res Function(_WeatherReportDto) _then) =
-      __$WeatherReportDtoCopyWithImpl;
+  factory _$WeatherReportSimpleDtoCopyWith(_WeatherReportSimpleDto value,
+          $Res Function(_WeatherReportSimpleDto) _then) =
+      __$WeatherReportSimpleDtoCopyWithImpl;
   @override
   @useResult
   $Res call(
-      {String reportId,
+      {String id,
       String userId,
-      String userName,
-      String mainPhotoUrl,
       String comment,
-      WeatherType weatherType,
-      FeelingType feelingType,
-      ForecastType forecastType,
       MediaType mediaType,
       String url,
-      DateTime createdAt,
-      List<String>? reportComment,
-      int? likeCount,
-      int? commentCount,
-      double? latitude,
-      double? longitude});
+      @IsoDateTimeConverter() DateTime createdAt});
 }
 
 /// @nodoc
-class __$WeatherReportDtoCopyWithImpl<$Res>
-    implements _$WeatherReportDtoCopyWith<$Res> {
-  __$WeatherReportDtoCopyWithImpl(this._self, this._then);
+class __$WeatherReportSimpleDtoCopyWithImpl<$Res>
+    implements _$WeatherReportSimpleDtoCopyWith<$Res> {
+  __$WeatherReportSimpleDtoCopyWithImpl(this._self, this._then);
 
-  final _WeatherReportDto _self;
-  final $Res Function(_WeatherReportDto) _then;
+  final _WeatherReportSimpleDto _self;
+  final $Res Function(_WeatherReportSimpleDto) _then;
 
   /// Create a copy of WeatherReportDto
   /// with the given fields replaced by the non-null parameter values.
   @override
   @pragma('vm:prefer-inline')
   $Res call({
-    Object? reportId = null,
+    Object? id = null,
     Object? userId = null,
-    Object? userName = null,
-    Object? mainPhotoUrl = null,
     Object? comment = null,
-    Object? weatherType = null,
-    Object? feelingType = null,
-    Object? forecastType = null,
     Object? mediaType = null,
     Object? url = null,
     Object? createdAt = null,
-    Object? reportComment = freezed,
-    Object? likeCount = freezed,
-    Object? commentCount = freezed,
-    Object? latitude = freezed,
-    Object? longitude = freezed,
   }) {
-    return _then(_WeatherReportDto(
-      reportId: null == reportId
-          ? _self.reportId
-          : reportId // ignore: cast_nullable_to_non_nullable
+    return _then(_WeatherReportSimpleDto(
+      id: null == id
+          ? _self.id
+          : id // ignore: cast_nullable_to_non_nullable
               as String,
       userId: null == userId
           ? _self.userId
           : userId // ignore: cast_nullable_to_non_nullable
               as String,
-      userName: null == userName
-          ? _self.userName
-          : userName // ignore: cast_nullable_to_non_nullable
-              as String,
-      mainPhotoUrl: null == mainPhotoUrl
-          ? _self.mainPhotoUrl
-          : mainPhotoUrl // ignore: cast_nullable_to_non_nullable
-              as String,
       comment: null == comment
           ? _self.comment
           : comment // ignore: cast_nullable_to_non_nullable
               as String,
-      weatherType: null == weatherType
-          ? _self.weatherType
-          : weatherType // ignore: cast_nullable_to_non_nullable
-              as WeatherType,
-      feelingType: null == feelingType
-          ? _self.feelingType
-          : feelingType // ignore: cast_nullable_to_non_nullable
-              as FeelingType,
-      forecastType: null == forecastType
-          ? _self.forecastType
-          : forecastType // ignore: cast_nullable_to_non_nullable
-              as ForecastType,
       mediaType: null == mediaType
           ? _self.mediaType
           : mediaType // ignore: cast_nullable_to_non_nullable
@@ -730,26 +438,6 @@ class __$WeatherReportDtoCopyWithImpl<$Res>
           ? _self.createdAt
           : createdAt // ignore: cast_nullable_to_non_nullable
               as DateTime,
-      reportComment: freezed == reportComment
-          ? _self._reportComment
-          : reportComment // ignore: cast_nullable_to_non_nullable
-              as List<String>?,
-      likeCount: freezed == likeCount
-          ? _self.likeCount
-          : likeCount // ignore: cast_nullable_to_non_nullable
-              as int?,
-      commentCount: freezed == commentCount
-          ? _self.commentCount
-          : commentCount // ignore: cast_nullable_to_non_nullable
-              as int?,
-      latitude: freezed == latitude
-          ? _self.latitude
-          : latitude // ignore: cast_nullable_to_non_nullable
-              as double?,
-      longitude: freezed == longitude
-          ? _self.longitude
-          : longitude // ignore: cast_nullable_to_non_nullable
-              as double?,
     ));
   }
 }
